rename project to mpvremote
