+++ MpvRemote/app/src/androidTest/java/miccah/mpvremote/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package miccah.laptopremote;
+package miccah.mpvremote;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("miccah.laptopremote", appContext.getPackageName());
+        assertEquals("miccah.mpvremote", appContext.getPackageName());
     }
 }
