+++ MpvRemote/app/src/main/java/miccah/mpvremote/HMAC.java
