+++ MpvRemote/app/src/main/java/miccah/mpvremote/Settings.java
