+++ MpvRemote/app/src/main/java/miccah/mpvremote/SettingsAdapter.java
