+++ MpvRemote/app/src/test/java/miccah/mpvremote/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package miccah.laptopremote;
+package miccah.mpvremote;
 
 import org.junit.Test;
